+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 * fix(Input|Textarea): padding for `noPadding+loading`
 * fix(components): improve `disabled` state
+* fix(CommandPalette): improve `back` button and divide color
 
 ## [2.4.1](https://github.com/bitrix24/b24ui/compare/v2.4.0...v2.4.1) (2026-03-04)